Flatten DTO packages into single namespace
