+++ src/main/java/io/github/ahumadamob/plangastos/dto/CuentaFinancieraResponseDto.java
@@ -1,4 +1,4 @@
-package io.github.ahumadamob.plangastos.dto.cuentafinanciera;
+package io.github.ahumadamob.plangastos.dto;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
