+++ src/main/java/io/github/ahumadamob/plangastos/dto/PartidaPlanificadaRequestDto.java
@@ -1,30 +1,18 @@
-package io.github.ahumadamob.plangastos.dto.partidaplanificada;
+package io.github.ahumadamob.plangastos.dto;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 import io.github.ahumadamob.plangastos.entity.Presupuesto;
 import io.github.ahumadamob.plangastos.entity.Rubro;
 
-public class PartidaPlanificadaResponseDto {
+public class PartidaPlanificadaRequestDto {
 
-    private Long id;
     private Presupuesto presupuesto;
     private Rubro rubro;
     private String descripcion;
     private BigDecimal montoComprometido;
     private LocalDate fechaObjetivo;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
 
     public Presupuesto getPresupuesto() {
         return presupuesto;
@@ -64,21 +52,5 @@ public class PartidaPlanificadaResponseDto {
 
     public void setFechaObjetivo(LocalDate fechaObjetivo) {
         this.fechaObjetivo = fechaObjetivo;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(LocalDateTime updatedAt) {
-        this.updatedAt = updatedAt;
     }
 }
