+++ src/main/java/io/github/ahumadamob/plangastos/dto/PlanPresupuestarioRequestDto.java
@@ -1,15 +1,14 @@
-package io.github.ahumadamob.plangastos.dto.rubro;
+package io.github.ahumadamob.plangastos.dto;
 
-import io.github.ahumadamob.plangastos.entity.NaturalezaMovimiento;
-import io.github.ahumadamob.plangastos.entity.Rubro;
+import io.github.ahumadamob.plangastos.entity.Divisa;
 import io.github.ahumadamob.plangastos.entity.Usuario;
 
-public class RubroRequestDto {
+public class PlanPresupuestarioRequestDto {
 
     private Usuario usuario;
-    private NaturalezaMovimiento naturaleza;
+    private Divisa divisa;
     private String nombre;
-    private Rubro parent;
+    private String descripcion;
     private Boolean activo;
 
     public Usuario getUsuario() {
@@ -20,12 +19,12 @@ public class RubroRequestDto {
         this.usuario = usuario;
     }
 
-    public NaturalezaMovimiento getNaturaleza() {
-        return naturaleza;
+    public Divisa getDivisa() {
+        return divisa;
     }
 
-    public void setNaturaleza(NaturalezaMovimiento naturaleza) {
-        this.naturaleza = naturaleza;
+    public void setDivisa(Divisa divisa) {
+        this.divisa = divisa;
     }
 
     public String getNombre() {
@@ -36,12 +35,12 @@ public class RubroRequestDto {
         this.nombre = nombre;
     }
 
-    public Rubro getParent() {
-        return parent;
+    public String getDescripcion() {
+        return descripcion;
     }
 
-    public void setParent(Rubro parent) {
-        this.parent = parent;
+    public void setDescripcion(String descripcion) {
+        this.descripcion = descripcion;
     }
 
     public Boolean getActivo() {
