+++ src/main/java/io/github/ahumadamob/plangastos/dto/RubroResponseDto.java
@@ -1,17 +1,18 @@
-package io.github.ahumadamob.plangastos.dto.planpresupuestario;
+package io.github.ahumadamob.plangastos.dto;
 
 import java.time.LocalDateTime;
 
-import io.github.ahumadamob.plangastos.entity.Divisa;
+import io.github.ahumadamob.plangastos.entity.NaturalezaMovimiento;
+import io.github.ahumadamob.plangastos.entity.Rubro;
 import io.github.ahumadamob.plangastos.entity.Usuario;
 
-public class PlanPresupuestarioResponseDto {
+public class RubroResponseDto {
 
     private Long id;
     private Usuario usuario;
-    private Divisa divisa;
+    private NaturalezaMovimiento naturaleza;
     private String nombre;
-    private String descripcion;
+    private Rubro parent;
     private Boolean activo;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
@@ -32,12 +33,12 @@ public class PlanPresupuestarioResponseDto {
         this.usuario = usuario;
     }
 
-    public Divisa getDivisa() {
-        return divisa;
+    public NaturalezaMovimiento getNaturaleza() {
+        return naturaleza;
     }
 
-    public void setDivisa(Divisa divisa) {
-        this.divisa = divisa;
+    public void setNaturaleza(NaturalezaMovimiento naturaleza) {
+        this.naturaleza = naturaleza;
     }
 
     public String getNombre() {
@@ -48,12 +49,12 @@ public class PlanPresupuestarioResponseDto {
         this.nombre = nombre;
     }
 
-    public String getDescripcion() {
-        return descripcion;
+    public Rubro getParent() {
+        return parent;
     }
 
-    public void setDescripcion(String descripcion) {
-        this.descripcion = descripcion;
+    public void setParent(Rubro parent) {
+        this.parent = parent;
     }
 
     public Boolean getActivo() {
